refactro RPi temperature sensor

+++ config/boards/rpi/flash.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ "$EUID" -ne 0 ]
+  then echo "ERROR: Please run as root"
+  exit
+fi
+
 pushd /home/klipper/klipper
 systemctl stop klipper.service
 systemctl stop klipper-mcu.service
